% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREprecompDb.R
\docType{data}
\name{CENTREprecompDb}
\alias{CENTREprecompDb}
\alias{CENTREprecompDb-class}
\alias{show}
\alias{dbconn,CENTREprecompDb-method}
\alias{show,CENTREprecompDb-method}
\title{Database for the CENTRE precomputed data}
\format{
An object of class \code{CENTREprecompDb} of length 1.
}
\usage{
CENTREprecompDb
}
\description{
\code{CENTREprecompDb} objects provide access to general (metabolite) compound
annotations along with \emph{metadata} information such as the annotation's
source, date and release version. The data is stored internally in a
database (usually an SQLite database).
}
\details{
\code{CENTREprecompDb} objects should be created using the constructor function
\code{CENTREprecompDb()} providing the name of the (SQLite) database file providing
the compound annotation data.
}
\keyword{datasets}
