\docType{class}
\name{CENTREprecomputedDb-class}
\alias{CENTREprecomputedDb-class}
\alias{.CENTREprecomputedDb}
\title{The CENTREprecomputedDb S4 class}
\value{
}
\description{
The CENTREprecomputedDb class is a subclass of the AnnotationDb class defined in the AnnotationDbi package. It is designed to hold annotations for CENTREprecomputed data.
}

