% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryEngine.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from the CENTREprecompDb database}
\usage{
fetch_data(x, table, columns, filter, column_filter)
}
\arguments{
\item{x}{\code{CENTREprecompDb} object.}

\item{table}{Table of the database on which to search. Can be combinedTestData
or cor_CRUP}

\item{columns}{Columns to select. Equivalent to X in SELECT X.}

\item{filter}{Element ID to select. Equivalen to ID in SELECT X FROM TABLE
WHERE ID in Z. If filter or column_filter is missing the program assumes the
query is SELECT X FROM TABLE.}

\item{column_filter}{Column on which to apply filter. Equivalent to Z in
SELECT X FROM TABLE WHERE ID in Z. If filter or column_filter is missing the
program assumes the query is SELECT X FROM TABLE.}
}
\description{
Main interface to fetch data from the CENTREprecomputed package database
through the \code{CENTREprecompDb} object.
}
\examples{
  res <- fetch_data(CENTREannotDb, 
                    columns=c("enhancer_id", "start"), 
                    filter=c("EH38E1519134","EH38E1519132") , 
                    column_filter="enhancer_id")
}
\references{
Based on \link{CompoundDb} query engine internal functions.
}
\seealso{
\code{vignette("CENTREprecomputed")}
}
