% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREprecompDb.R
\docType{class}
\name{CENTREprecompDb}
\alias{CENTREprecompDb}
\alias{CENTREprecompDb-class}
\alias{show}
\alias{dbconn}
\alias{tables}
\title{Database for the CENTRE precomputed data}
\usage{
CENTREprecompDb(x)

tables(x)
}
\arguments{
\item{x}{sqlite file path}
}
\value{
Object of class \code{CENTREprecompDb}
}
\description{
The \code{CENTREprecompDb} object provides access to CENTRE's Precomputed SQLite
Database PrecomputedDataLight.db. Inside the database is combinedTestData,
crup_cor and metadata. For more information check the
computeGenericFeatures() function from the CENTRE package.
}
\details{
\code{CENTREprecompDb} is the object that provides access to CENTRE's database.
\code{tables(CENTREprecompDb)} shows the tables inside the database and
their columns.
}
\examples{
#The object is accessed through the ExperimentHub record
eh <- ExperimentHub::ExperimentHub()
centreprecompdb <- eh[["EH9540"]]
tables(centreprecompdb)

}
\references{
Based on CompoundDb::CompDb class.
}
