% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREprecompDb.R
\docType{class}
\name{CENTREprecompDb}
\alias{CENTREprecompDb}
\alias{CENTREprecompDb-class}
\alias{show}
\alias{dbconn}
\alias{CENTREprecomputedDb}
\alias{tables}
\title{Database for the CENTRE precomputed data}
\usage{
CENTREprecomputedDb(x)

tables(x)
}
\arguments{
\item{x}{sqlite file path}
}
\description{
The \code{CENTREprecompDb} object provides access to CENTRE's Precomputed SQLite
Database PrecomputedDataLight.db. Inside the database is combinedTestData,
crup_cor and metadata. For more information check the computeGenericFeatures()
function from the CENTRE package
}
\details{
\code{CENTREprecompDb} is the object that provides access to CENTRE's database.
\code{CENTREprecompDb@.properties$tables} shows the tables inside the database and
their columns.
}
\references{
Based on \link[CompoundDb:CompDb]{CompoundDb::CompDb} class.
}
