% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREprecompDb.R
\docType{class}
\name{CENTREprecompDb}
\alias{CENTREprecompDb}
\alias{CENTREprecompDb-class}
\alias{show}
\alias{dbconn,CENTREprecompDb-method}
\alias{show,CENTREprecompDb-method}
\alias{tables}
\title{Database for the CENTRE precomputed data}
\format{
An object of class \code{CENTREprecompDb} of length 1.
}
\usage{
CENTREprecompDb

tables(x)
}
\description{
The \code{CENTREprecompDb} object provides access to CENTRE's Precomputed SQLite
Database PrecomputedDataLight.db. Inside the database is combinedTestData,
crup_cor and metadata. For more information check \link[CENTRE]{computeGenericFeatures}
}
\details{
\code{CENTREprecompDb} is the object that provides access to CENTRE's database.
\code{CENTREprecompDb@.properties$tables} shows the tables inside the database and
their columns.
}
\keyword{datasets}
